Move InfoIcon to shared component file
