+++ frontend/src/metabase/components/MetadataInfo/InfoIcon/index.tsx
@@ -1,0 +1,1 @@
+export * from "./InfoIcon.styled";
